fix(mattermost): keep lifecycle status helper private
